requirements.txt
python-dotenv>=1.0
mss>=9.0
Pillow>=10.0
opencv-python>=4.8
pytesseract>=0.3.10
easyocr>=1.7
anthropic>=0.20
openai>=1.0
httpx>=0.25
PyQt6>=6.6
pynput>=1.7
keyring>=24.0
pydantic>=2.0
PyYAML>=6.0
loguru>=0.7
pytest>=7.0
pytest-asyncio>=0.23
